move select to AssignHearingsContainer, all calls handled by LoadingDataDisplay
