+++ client/app/hearingSchedule/containers/AssignHearingsContainer.jsx
@@ -2,32 +2,54 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
+import COPY from '../../../COPY.json';
 import _ from 'lodash';
-import { LOGO_COLORS } from '../../constants/AppConstants';
+import { css } from 'glamor';
+import AppSegment from '@department-of-veterans-affairs/caseflow-frontend-toolkit/components/AppSegment';
+import Link from '@department-of-veterans-affairs/caseflow-frontend-toolkit/components/Link';
 import ApiUtil from '../../util/ApiUtil';
-import LoadingDataDisplay from '../../components/LoadingDataDisplay';
 import {
   onRegionalOfficeChange,
   onReceiveUpcomingHearingDays,
   onSelectedHearingDayChange,
   onReceiveVeteransReadyForHearing
 } from '../actions';
+import LoadingDataDisplay from '../../components/LoadingDataDisplay';
+import { COLORS, LOGO_COLORS } from '../../constants/AppConstants';
 import { onReceiveTasks } from '../../queue/QueueActions';
 import { setUserCssId } from '../../queue/uiReducer/uiActions';
+import RoSelectorDropdown from '../components/RoSelectorDropdown';
 import AssignHearings from '../components/AssignHearings';
+
+const centralOfficeStaticEntry = [{
+  label: 'Central',
+  value: 'C'
+}];
+
+const smallTopMargin = css({
+  fontStyle: 'italic',
+  '.usa-input-error': {
+    marginTop: '1rem'
+  },
+  '.usa-input-error-message': {
+    paddingBottom: '0',
+    paddingTop: '0',
+    right: '0'
+  },
+  '& > p': {
+    fontWeight: '500',
+    color: COLORS.RED_DARK,
+    marginBottom: '0',
+    fontSize: '1.7rem',
+    marginTop: '1px'
+  }
+});
 
 class AssignHearingsContainer extends React.PureComponent {
 
   componentDidMount = () => {
     this.props.setUserCssId(this.props.userCssId);
   }
-
-  componentDidUpdate = (prevProps) => {
-    if (this.props.selectedRegionalOffice !== prevProps.selectedRegionalOffice) {
-      this.loadUpcomingHearingDays();
-      this.loadVeteransReadyForHearing();
-    }
-  };
 
   loadUpcomingHearingDays = () => {
     if (!this.props.selectedRegionalOffice) {
@@ -60,30 +82,57 @@ class AssignHearingsContainer extends React.PureComponent {
     });
   };
 
-  createLoadPromise = () => Promise.all([
-    this.loadUpcomingHearingDays()
-  ]);
+  getNoUpcomingError = () => {
+    if (this.props.selectedRegionalOffice) {
+      return <div className="usa-input-error-message usa-input-error" {...smallTopMargin}>
+        <span>{this.props.selectedRegionalOffice && this.props.selectedRegionalOffice.label} has
+          no upcoming hearing days.</span><br />
+        <p>Please verify that this RO's hearing days are in the current schedule.</p>
+      </div>;
+    }
+  }
+
+  createLoadPromise = () => {
+    return Promise.all([
+      this.loadUpcomingHearingDays(), this.loadVeteransReadyForHearing()
+    ]);
+  }
 
   render = () => {
-    const loadingDataDisplay = <LoadingDataDisplay
-      createLoadPromise={this.createLoadPromise}
-      loadingComponentProps={{
-        spinnerColor: LOGO_COLORS.HEARING_SCHEDULE.ACCENT,
-        message: 'Loading appeals to be scheduled for hearings...'
-      }}>
-      <AssignHearings
-        onRegionalOfficeChange={this.props.onRegionalOfficeChange}
-        selectedRegionalOffice={this.props.selectedRegionalOffice}
-        upcomingHearingDays={this.props.upcomingHearingDays}
-        onSelectedHearingDayChange={this.props.onSelectedHearingDayChange}
-        selectedHearingDay={this.props.selectedHearingDay}
-        veteransReadyForHearing={this.props.veteransReadyForHearing}
-        userId={this.props.userId}
-        onReceiveTasks={this.props.onReceiveTasks}
-      />
-    </LoadingDataDisplay>;
+    return (
+      <AppSegment filledBackground>
+        <h1>{COPY.HEARING_SCHEDULE_ASSIGN_HEARINGS_HEADER}</h1>
+        <Link
+          name="view-schedule"
+          to="/schedule">
+          {COPY.HEARING_SCHEDULE_ASSIGN_HEARINGS_VIEW_SCHEDULE_LINK}
+        </Link>
+        <RoSelectorDropdown
+          onChange={this.props.onRegionalOfficeChange}
+          value={this.props.selectedRegionalOffice}
+          staticOptions={centralOfficeStaticEntry}
+        />
+        {this.props.selectedRegionalOffice && <LoadingDataDisplay
+          key={this.props.selectedRegionalOffice.value}
+          createLoadPromise={this.createLoadPromise}
+          loadingComponentProps={{
+            spinnerColor: LOGO_COLORS.HEARING_SCHEDULE.ACCENT,
+            message: 'Loading appeals to be scheduled for hearings...'
+          }}>
 
-    return <div>{loadingDataDisplay}</div>;
+          {_.isEmpty(this.props.upcomingHearingDays) && this.getNoUpcomingError()}
+
+          <AssignHearings
+            selectedRegionalOffice={this.props.selectedRegionalOffice}
+            upcomingHearingDays={this.props.upcomingHearingDays}
+            onSelectedHearingDayChange={this.props.onSelectedHearingDayChange}
+            selectedHearingDay={this.props.selectedHearingDay}
+            veteransReadyForHearing={this.props.veteransReadyForHearing}
+            userId={this.props.userId}
+            onReceiveTasks={this.props.onReceiveTasks} />
+        </LoadingDataDisplay>}
+      </AppSegment>
+    );
   }
 }
 
@@ -94,6 +143,7 @@ AssignHearings.propTypes = {
 
 const mapStateToProps = (state) => ({
   selectedRegionalOffice: state.hearingSchedule.selectedRegionalOffice,
+  fetchingHearingDays: state.hearingSchedule.fetchingHearingDays,
   upcomingHearingDays: state.hearingSchedule.upcomingHearingDays,
   selectedHearingDay: state.hearingSchedule.selectedHearingDay,
   veteransReadyForHearing: state.hearingSchedule.veteransReadyForHearing
